backend/requirements.txt
fastapi
uvicorn
PyJWT
python-multipart
supabase
pydantic
pydantic-settings
python-dotenv
openpyxl
PyPDF2
httpx
